Add files via upload
Every day, hot dog vendors at Metlife Stadium are randomly assigned sections of the stadium for which they are responsible. The data shows for each game and section of the stadium the vendor assigned to that area, the day of the week (1 = Sunday, 2 = Monday, etc.), and the number of hot dogs that were sold. Developing a model that can be used to rank vendors’ sales abilities.